Apply a hotfix for flakey DPDK parallel build
DPDK has a race between building C files and creating symbolic links
(`lib/librte_ethdev/*.h` -> `build/include/*.h`) for their dependency
headers. This race may cause a build failure on `make -j`. While it
is fairly rare, it tends to happen more often on busy machines.

The easiest way to reproduce the bug is to add an artificial delay in
the Makefile:

```
--- a/mk/internal/rte.install-post.mk
+++ b/mk/internal/rte.install-post.mk
@@ -31,6 +31,7 @@ define symlink_rule
 $(addprefix $(RTE_OUTPUT)/$(1)/,$(notdir $(2))): $(2)
        @echo "  SYMLINK-FILE $(addprefix $(1)/,$(notdir $(2)))"
        @[ -d $(RTE_OUTPUT)/$(1) ] || mkdir -p $(RTE_OUTPUT)/$(1)
+       sleep 10
        $(Q)ln -nsf `$(RTE_SDK)/buildtools/relpath.sh $$(<) $(RTE_OUTPUT)/$(1)` \
                $(RTE_OUTPUT)/$(1)
 endef
```

I am going to send a patch to the DPDK upstream, after figurng out
what the root cause of the race is. Until then, we work around the
problem with a patch.