Add metadata with attribute to the Timestamp module.
This patch is useful when the Timestamp and Measure modules are
used in tandem. In current version, the default location to store
the metadata in Timestamp module is at Eth/IP4/Udp offset. This
may lead to bugs if the pipeline uses encap/decap operations
(header sizes are changed, and timestamp value can get corrupted).

We propose using attribute option to store timestamp value as
default.

Signed-off-by: Muhammad Asim Jamshed <muhammad.jamshed@intel.com>
Signed-off-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>