Relax naming restrictions in bessctl commands
Currently, in bessctl, names of modules, traffic classes, etc. have to
be compliant to Python identifier: [_a-zA-Z][_a-zA-Z0-9]*
This design was due to the plan for tight integration (one to one
mapping) between Python variable names and BESS module names, but this
has never been implemented.

Instead, this commit eliminates the naming restrictions in bessctl
commands. For example, now module names can be specified as a number,
URL, unicode string, etc. Whitespaces are still not allowed (this is
partly because of a parsing issue; bessctl commands do not support
escape characters at the moment), but it may change in the future.

Note that there still are limitations regarding names defined in bessctl
script with "syntax sugar", e.g., `foo::Bar()`, where `foo` must be a
legal Python variable name.