RPF (no reflection) check
This is a set of basic functionality which is present in all
switching and forwarding frameworks. They all set/keep packet
origin in the packet metadata.

The reason is that knowing your "in" interface is essential
for most forwarding decisions. Example - you cannot implement
even a basic hub without knowing your "in" interface, because
you are not supposed to send packets to the interface which
originated them. There are similar requirements for forwarding,
security, acls, etc.

This changeset adds the following functionality:

1. Each port gets an ifIndex similar to ifIndex in SNMP assigned
   on creation.

2. PortInc now always sets an integer metadata called ifIndex
   with the ifIndex value.

3. PortOut now can optionally check if the packet is being sent
   out from its arrival port and drop it.

The default behavious is rpf check off - to emulate previous
BESS behaviour.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>